+++ src/main/java/frc/robot/Robot.java
@@ -80,7 +80,7 @@ public class Robot extends LoggedRobot {
 
     // Version Number (# of Pushes to Dev, Issue #, Commit #, Functionality 0 = working, 1 = WIP, 2
     // = doesn't work)
-    SmartDashboard.putString("Version Number", "10.0.10.0");
+    SmartDashboard.putString("Version Number", "10.5.11.1");
     SmartDashboard.putString("Last Deployed: ", BuildConstants.BUILD_DATE);
 
     m_robotContainer = new RobotContainer();
